+++ Monofoxe/Monofoxe.Engine/Drawing/Frame.cs
@@ -23,16 +23,9 @@ namespace Monofoxe.Engine.Drawing
 		public readonly RectangleF TexturePosition;
 
 		/// <summary>
-		/// Width of the frame.
+		/// Size of the frame.
 		/// </summary>
-		public float Width => TexturePosition.Width;
-
-		/// <summary>
-		/// Height of the frame.
-		/// </summary>
-		public float Height => TexturePosition.Height;
-		
-
+		public Vector2 Size => TexturePosition.Size;
 		
 		public Vector2 Scale = Vector2.One;
 		
@@ -113,14 +106,14 @@ namespace Monofoxe.Engine.Drawing
 			{
 				flipFlags = flipFlags | SpriteFlipFlags.FlipHorizontally;
 				scale.X *= -1;
-				origin.X = Width - origin.X;
+				origin.X = Size.X - origin.X;
 			}
 
 			if (scale.Y < 0)
 			{
 				flipFlags = flipFlags | SpriteFlipFlags.FlipVertically;
 				scale.Y *= -1;
-				origin.Y = Height - origin.Y;
+				origin.Y = Size.Y - origin.Y;
 			}
 			// Proper negative scaling.