+++ Samples/Monofoxe.Samples/Demos/ShapeDemo.cs
@@ -38,10 +38,10 @@ namespace Monofoxe.Samples.Demos
 
 			// Circles.
 			GraphicsMgr.CurrentColor = _mainColor; // Setting current color. It's active for all shapes and sprites.
-			CircleShape.Draw(position, 24, false); // Filled circle.
+			CircleShape.Draw(position, 24, ShapeFill.Solid); // Filled circle.
 
 			GraphicsMgr.CurrentColor = _secondaryColor.ToColor();
-			CircleShape.Draw(position, 32, true); // Outline.
+			CircleShape.Draw(position, 32, ShapeFill.Outline); // Outline.
 
 
 			position += Vector2.UnitX * spacing;
@@ -50,10 +50,10 @@ namespace Monofoxe.Samples.Demos
 			CircleShape.CircleVerticesCount = 8; // Changing the amount of circle vertices.
 
 			GraphicsMgr.CurrentColor = _mainColor; 
-			CircleShape.Draw(position, 24, false); 
+			CircleShape.Draw(position, 24, ShapeFill.Solid); 
 
 			GraphicsMgr.CurrentColor = _secondaryColor.ToColor();
-			CircleShape.Draw(position, 32, true);
+			CircleShape.Draw(position, 32, ShapeFill.Outline);
 
 			CircleShape.CircleVerticesCount = 32;
 			// Circles.
@@ -66,18 +66,18 @@ namespace Monofoxe.Samples.Demos
 			// Rectangles.
 			
 			// You can draw rectangle using its top left and bottom right point...
-			RectangleShape.Draw(position - Vector2.One * 24, position + Vector2.One * 24, false);
+			RectangleShape.Draw(position - Vector2.One * 24, position + Vector2.One * 24, ShapeFill.Solid);
 
 			GraphicsMgr.CurrentColor = _mainColor;
 			// ...or its center position and size!
-			RectangleShape.DrawBySize(position, Vector2.One * 64, true);
+			RectangleShape.DrawBySize(position, Vector2.One * 64, ShapeFill.Outline);
 
 			position += Vector2.UnitX * spacing;
 
 			RectangleShape.Draw( // We can also manually set colors for each vertex.
 				position - Vector2.One * 24, 
-				position + Vector2.One * 24, 
-				false, 
+				position + Vector2.One * 24,
+				ShapeFill.Solid, 
 				_mainColor, 
 				_mainColor, 
 				_mainColor, 
@@ -86,8 +86,8 @@ namespace Monofoxe.Samples.Demos
 
 			RectangleShape.DrawBySize(
 				position, 
-				Vector2.One * 64, 
-				true,
+				Vector2.One * 64,
+				ShapeFill.Outline,
 				_mainColor,
 				_secondaryColor.ToColor(),
 				_mainColor,
@@ -107,13 +107,13 @@ namespace Monofoxe.Samples.Demos
 			TriangleShape.Draw(
 				position + new Vector2(-24, -24), 
 				position + new Vector2(24, -24), 
-				position + new Vector2(24, 24), 
-				false
+				position + new Vector2(24, 24),
+				ShapeFill.Solid
 			);
 			
 			// Be aware of culling. This triangle, for example, will be culled.
 			// You can disable culling, if you don't want to deal with it.
-			TriangleShape.Draw(new Vector2(-24, -24), new Vector2(24, 24), new Vector2(24, -24), false);
+			TriangleShape.Draw(new Vector2(-24, -24), new Vector2(24, 24), new Vector2(24, -24), ShapeFill.Solid);
 
 			// Triangles.