+++ Samples/Monofoxe.Samples/Misc/Ball.cs
@@ -46,7 +46,7 @@ namespace Monofoxe.Samples.Misc
 		{
 			base.Draw();
 			GraphicsMgr.CurrentColor = _color;
-			CircleShape.Draw(_position, _r, false);
+			CircleShape.Draw(_position, _r, ShapeFill.Solid);
 		}
 
 	}